requirements.txt
cryptography
dateparser
prettytable
urllib3
sniffio
python-telegram-bot
forex-python
asyncio